refactor:former unit test for the UserTest Entity have been adapt to adapt to functional test for the UserControllerTest and test-coverage have been removed temporarily
