+++ .gitignore
@@ -36,7 +36,5 @@ phpstan.neon
 
 grumphp.yml
 
-###> phpunit/phpunit ###
-/phpunit.xml
-.phpunit.result.cache
-###< phpunit/phpunit ###
+
+/public/test-coverage
