--- public/test-coverage/DataFixtures/index.html
@@ -1,89 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
- <head>
-  <meta charset="UTF-8">
-  <title>Code Coverage for C:\xampp\htdocs\TodoList\src\DataFixtures</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link href="../_css/bootstrap.min.css?v=9.2.29" rel="stylesheet" type="text/css">
-  <link href="../_css/octicons.css?v=9.2.29" rel="stylesheet" type="text/css">
-  <link href="../_css/style.css?v=9.2.29" rel="stylesheet" type="text/css">
-  <link href="../_css/custom.css" rel="stylesheet" type="text/css">
- </head>
- <body>
-  <header>
-   <div class="container-fluid">
-    <div class="row">
-     <div class="col-md-12">
-      <nav aria-label="breadcrumb">
-       <ol class="breadcrumb">
-         <li class="breadcrumb-item"><a href="../index.html">C:\xampp\htdocs\TodoList\src</a></li>
-         <li class="breadcrumb-item active">DataFixtures</li>
-         <li class="breadcrumb-item">(<a href="dashboard.html">Dashboard</a>)</li>
-
-       </ol>
-      </nav>
-     </div>
-    </div>
-   </div>
-  </header>
-  <div class="container-fluid">
-   <div class="table-responsive">
-    <table class="table table-bordered">
-     <thead>
-      <tr>
-       <td>&nbsp;</td>
-       <td colspan="9"><div align="center"><strong>Code Coverage</strong></div></td>
-      </tr>
-      <tr>
-       <td>&nbsp;</td>
-       <td colspan="3"><div align="center"><strong>Lines</strong></div></td>
-       <td colspan="3"><div align="center"><strong>Functions and Methods</strong></div></td>
-       <td colspan="3"><div align="center"><strong>Classes and Traits</strong></div></td>
-      </tr>
-     </thead>
-     <tbody>
-      <tr>
-       <td class="">Total</td>
-       <td class=" big"></td>
-       <td class=" small"><div align="right">n/a</div></td>
-       <td class=" small"><div align="right">0&nbsp;/&nbsp;0</div></td>
-       <td class=" big"></td>
-       <td class=" small"><div align="right">n/a</div></td>
-       <td class=" small"><div align="right">0&nbsp;/&nbsp;0</div></td>
-       <td class=" big"></td>
-       <td class=" small"><div align="right">n/a</div></td>
-       <td class=" small"><div align="right">0&nbsp;/&nbsp;0</div></td>
-      </tr>
-
-      <tr>
-       <td class=""><img src="../_icons/file-code.svg" class="octicon" /><a href="AppFixtures.php.html">AppFixtures.php</a></td>
-       <td class=" big"></td>
-       <td class=" small"><div align="right">n/a</div></td>
-       <td class=" small"><div align="right">0&nbsp;/&nbsp;0</div></td>
-       <td class=" big"></td>
-       <td class=" small"><div align="right">n/a</div></td>
-       <td class=" small"><div align="right">0&nbsp;/&nbsp;0</div></td>
-       <td class=" big"></td>
-       <td class=" small"><div align="right">n/a</div></td>
-       <td class=" small"><div align="right">0&nbsp;/&nbsp;0</div></td>
-      </tr>
-
-
-     </tbody>
-    </table>
-   </div>
-   <footer>
-    <hr/>
-    <h4>Legend</h4>
-    <p>
-     <span class="danger"><strong>Low</strong>: 0% to 50%</span>
-     <span class="warning"><strong>Medium</strong>: 50% to 90%</span>
-     <span class="success"><strong>High</strong>: 90% to 100%</span>
-    </p>
-    <p>
-     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.29</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.12</a> and <a href="https://phpunit.de/">PHPUnit 9.6.15</a> at Sun Dec 10 17:18:01 CET 2023.</small>
-    </p>
-   </footer>
-  </div>
- </body>
-</html>
