--- public/test-coverage/Entity/index.html
@@ -1,147 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
- <head>
-  <meta charset="UTF-8">
-  <title>Code Coverage for C:\xampp\htdocs\TodoList\src\Entity</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link href="../_css/bootstrap.min.css?v=9.2.29" rel="stylesheet" type="text/css">
-  <link href="../_css/octicons.css?v=9.2.29" rel="stylesheet" type="text/css">
-  <link href="../_css/style.css?v=9.2.29" rel="stylesheet" type="text/css">
-  <link href="../_css/custom.css" rel="stylesheet" type="text/css">
- </head>
- <body>
-  <header>
-   <div class="container-fluid">
-    <div class="row">
-     <div class="col-md-12">
-      <nav aria-label="breadcrumb">
-       <ol class="breadcrumb">
-         <li class="breadcrumb-item"><a href="../index.html">C:\xampp\htdocs\TodoList\src</a></li>
-         <li class="breadcrumb-item active">Entity</li>
-         <li class="breadcrumb-item">(<a href="dashboard.html">Dashboard</a>)</li>
-
-       </ol>
-      </nav>
-     </div>
-    </div>
-   </div>
-  </header>
-  <div class="container-fluid">
-   <div class="table-responsive">
-    <table class="table table-bordered">
-     <thead>
-      <tr>
-       <td>&nbsp;</td>
-       <td colspan="9"><div align="center"><strong>Code Coverage</strong></div></td>
-      </tr>
-      <tr>
-       <td>&nbsp;</td>
-       <td colspan="3"><div align="center"><strong>Lines</strong></div></td>
-       <td colspan="3"><div align="center"><strong>Functions and Methods</strong></div></td>
-       <td colspan="3"><div align="center"><strong>Classes and Traits</strong></div></td>
-      </tr>
-     </thead>
-     <tbody>
-      <tr>
-       <td class="warning">Total</td>
-       <td class="warning big">       <div class="progress">
-         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="85.71" aria-valuemin="0" aria-valuemax="100" style="width: 85.71%">
-           <span class="sr-only">85.71% covered (warning)</span>
-         </div>
-       </div>
-</td>
-       <td class="warning small"><div align="right">85.71%</div></td>
-       <td class="warning small"><div align="right">24&nbsp;/&nbsp;28</div></td>
-       <td class="warning big">       <div class="progress">
-         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="84.62" aria-valuemin="0" aria-valuemax="100" style="width: 84.62%">
-           <span class="sr-only">84.62% covered (warning)</span>
-         </div>
-       </div>
-</td>
-       <td class="warning small"><div align="right">84.62%</div></td>
-       <td class="warning small"><div align="right">22&nbsp;/&nbsp;26</div></td>
-       <td class="danger big">       <div class="progress">
-         <div class="progress-bar bg-danger" role="progressbar" aria-valuenow="0.00" aria-valuemin="0" aria-valuemax="100" style="width: 0.00%">
-           <span class="sr-only">0.00% covered (danger)</span>
-         </div>
-       </div>
-</td>
-       <td class="danger small"><div align="right">0.00%</div></td>
-       <td class="danger small"><div align="right">0&nbsp;/&nbsp;2</div></td>
-      </tr>
-
-      <tr>
-       <td class="warning"><img src="../_icons/file-code.svg" class="octicon" /><a href="Task.php.html">Task.php</a></td>
-       <td class="warning big">       <div class="progress">
-         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="76.92" aria-valuemin="0" aria-valuemax="100" style="width: 76.92%">
-           <span class="sr-only">76.92% covered (warning)</span>
-         </div>
-       </div>
-</td>
-       <td class="warning small"><div align="right">76.92%</div></td>
-       <td class="warning small"><div align="right">10&nbsp;/&nbsp;13</div></td>
-       <td class="warning big">       <div class="progress">
-         <div class="progress-bar bg-warning" role="progressbar" aria-valuenow="75.00" aria-valuemin="0" aria-valuemax="100" style="width: 75.00%">
-           <span class="sr-only">75.00% covered (warning)</span>
-         </div>
-       </div>
-</td>
-       <td class="warning small"><div align="right">75.00%</div></td>
-       <td class="warning small"><div align="right">9&nbsp;/&nbsp;12</div></td>
-       <td class="danger big">       <div class="progress">
-         <div class="progress-bar bg-danger" role="progressbar" aria-valuenow="0.00" aria-valuemin="0" aria-valuemax="100" style="width: 0.00%">
-           <span class="sr-only">0.00% covered (danger)</span>
-         </div>
-       </div>
-</td>
-       <td class="danger small"><div align="right">0.00%</div></td>
-       <td class="danger small"><div align="right">0&nbsp;/&nbsp;1</div></td>
-      </tr>
-
-      <tr>
-       <td class="success"><img src="../_icons/file-code.svg" class="octicon" /><a href="User.php.html">User.php</a></td>
-       <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="93.33" aria-valuemin="0" aria-valuemax="100" style="width: 93.33%">
-           <span class="sr-only">93.33% covered (success)</span>
-         </div>
-       </div>
-</td>
-       <td class="success small"><div align="right">93.33%</div></td>
-       <td class="success small"><div align="right">14&nbsp;/&nbsp;15</div></td>
-       <td class="success big">       <div class="progress">
-         <div class="progress-bar bg-success" role="progressbar" aria-valuenow="92.86" aria-valuemin="0" aria-valuemax="100" style="width: 92.86%">
-           <span class="sr-only">92.86% covered (success)</span>
-         </div>
-       </div>
-</td>
-       <td class="success small"><div align="right">92.86%</div></td>
-       <td class="success small"><div align="right">13&nbsp;/&nbsp;14</div></td>
-       <td class="danger big">       <div class="progress">
-         <div class="progress-bar bg-danger" role="progressbar" aria-valuenow="0.00" aria-valuemin="0" aria-valuemax="100" style="width: 0.00%">
-           <span class="sr-only">0.00% covered (danger)</span>
-         </div>
-       </div>
-</td>
-       <td class="danger small"><div align="right">0.00%</div></td>
-       <td class="danger small"><div align="right">0&nbsp;/&nbsp;1</div></td>
-      </tr>
-
-
-     </tbody>
-    </table>
-   </div>
-   <footer>
-    <hr/>
-    <h4>Legend</h4>
-    <p>
-     <span class="danger"><strong>Low</strong>: 0% to 50%</span>
-     <span class="warning"><strong>Medium</strong>: 50% to 90%</span>
-     <span class="success"><strong>High</strong>: 90% to 100%</span>
-    </p>
-    <p>
-     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 9.2.29</a> using <a href="https://secure.php.net/" target="_top">PHP 8.2.12</a> and <a href="https://phpunit.de/">PHPUnit 9.6.15</a> at Sun Dec 10 17:18:01 CET 2023.</small>
-    </p>
-   </footer>
-  </div>
- </body>
-</html>
