--- public/test-coverage/_css/octicons.css
@@ -1,5 +1,0 @@
-.octicon {
-  display: inline-block;
-  vertical-align: text-top;
-  fill: currentColor;
-}
